♻️ Cria um RootStackNavigator
Melhora a organização das rotas raízes
