+++ App.tsx
@@ -1,19 +1,15 @@
 import React from "react";
 import AppLoading from "expo-app-loading";
 import { StatusBar } from "expo-status-bar";
-import OnBoarding from "./src/pages/OnBoarding";
 import { NavigationContainer } from "@react-navigation/native";
-import TopTabNavigator from "./src/pages/TopTabNavigator";
+import RootStackNavigator from "./src/navigators/RootStackNavigator";
 import { SafeAreaProvider } from "react-native-safe-area-context";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import {
     useFonts,
     Lexend_500Medium,
     Lexend_700Bold,
     Lexend_600SemiBold,
 } from "@expo-google-fonts/lexend";
-
-const Stack = createNativeStackNavigator();
 
 export default function App(): React.ReactElement {
     const [fontsLoaded] = useFonts({
@@ -30,18 +26,7 @@ export default function App(): React.ReactElement {
         <SafeAreaProvider>
             <NavigationContainer>
                 <StatusBar style="auto" />
-                <Stack.Navigator>
-                    <Stack.Screen
-                        name="OnBoarding"
-                        component={OnBoarding}
-                        options={{ headerShown: false }}
-                    />
-                    <Stack.Screen
-                        name="Home"
-                        component={TopTabNavigator}
-                        options={{ headerShown: false }}
-                    />
-                </Stack.Navigator>
+                <RootStackNavigator />
             </NavigationContainer>
         </SafeAreaProvider>
     );
